copy config from admin user for new users
