+++ views/auth.py
@@ -5,7 +5,7 @@ Authentication modules
 from functools import wraps
 from flask import session, request, jsonify
 from passlib.handlers.argon2 import argon2
-from db import User
+from db.model import User
 from views import application
 from views.postgres import get_db_session
