+++ views/gene.py
@@ -8,7 +8,7 @@ from views import application
 from views.auth import requires_auth
 from views.postgres import get_db_session, postgres_cursor
 from views.general import process_for_display
-from db import Gene
+from db.model import Gene
 
 
 @application.route("/<language>/gene/<gene_id>")
