+++ views/users.py
@@ -4,7 +4,7 @@ Users view
 from flask import session, request, jsonify
 from passlib.handlers.argon2 import argon2
 
-from db import User, User_Individual, Individual
+from db.model import User, UserIndividual, Individual, UserConfig
 from views import application
 from views.auth import requires_auth, check_auth
 from views.exceptions import PhenopolisException
@@ -67,21 +67,22 @@ def create_user():
     for u in new_users:
         u.argon_password = argon2.hash(u.argon_password)
 
-    sqlalchemy_session = get_db_session()
+    db_session = get_db_session()
     request_ok = True
     message = "Users were created"
     user_ids = ",".join([u.user for u in new_users])
     try:
         # insert users
-        sqlalchemy_session.add_all(new_users)
-        sqlalchemy_session.commit()
+        db_session.add_all(new_users)
+        _add_config_from_admin(db_session, new_users)
+        db_session.commit()
     except Exception as e:
-        sqlalchemy_session.rollback()
+        db_session.rollback()
         application.logger.exception(e)
         request_ok = False
         message = str(e)
     finally:
-        sqlalchemy_session.close()
+        db_session.close()
 
     if not request_ok:
         return jsonify(success=False, message=message), 500
@@ -102,7 +103,7 @@ def create_user_idividual():
 
     # parse the JSON data into a user_individual, non existing fields will trigger a TypeError
     try:
-        new_user_individuals = _parse_payload(payload, User_Individual)
+        new_user_individuals = _parse_payload(payload, UserIndividual)
     except TypeError as e:
         application.logger.error(str(e))
         return jsonify(success=False, error=str(e)), 400
@@ -122,21 +123,7 @@ def create_user_idividual():
         # insert user individuals
         for u in new_user_individuals:
             # TODO: should not all these checks happen at the DB?
-            if db_session.query(User.user).filter(User.user.match(u.user)).count() != 1:
-                raise PhenopolisException("Trying to add an entry in user_individual to a non existing user")
-            if (
-                db_session.query(Individual.internal_id).filter(Individual.internal_id.match(u.internal_id)).count()
-                != 1
-            ):
-                raise PhenopolisException("Trying to add an entry in user_individual to a non existing individual")
-            if (
-                db_session.query(User_Individual)
-                .filter(User_Individual.user.match(u.user))
-                .filter(User_Individual.internal_id.match(u.internal_id))
-                .count()
-                > 0
-            ):
-                raise PhenopolisException("Trying to add an entry in user_individual that already exists")
+            _check_db_integrity_user_individual(db_session, u)
             db_session.add(u)
         db_session.commit()
     except Exception as e:
@@ -153,6 +140,21 @@ def create_user_idividual():
         return jsonify(success=True, message=message), 200
 
 
+def _check_db_integrity_user_individual(db_session, u):
+    if db_session.query(User.user).filter(User.user.match(u.user)).count() != 1:
+        raise PhenopolisException("Trying to add an entry in user_individual to a non existing user")
+    if db_session.query(Individual.internal_id).filter(Individual.internal_id.match(u.internal_id)).count() != 1:
+        raise PhenopolisException("Trying to add an entry in user_individual to a non existing individual")
+    if (
+        db_session.query(UserIndividual)
+        .filter(UserIndividual.user.match(u.user))
+        .filter(UserIndividual.internal_id.match(u.internal_id))
+        .count()
+        > 0
+    ):
+        raise PhenopolisException("Trying to add an entry in user_individual that already exists")
+
+
 def _check_user_valid(new_user: User):
     if new_user is None:
         raise PhenopolisException("Null user")
@@ -162,10 +164,21 @@ def _check_user_valid(new_user: User):
         raise PhenopolisException("Missing password")
 
 
-def _check_user_individual_valid(new_user_individual: User_Individual):
+def _check_user_individual_valid(new_user_individual: UserIndividual):
     if new_user_individual is None:
         raise PhenopolisException("Null user individual")
     if new_user_individual.user is None or new_user_individual.user == "":
         raise PhenopolisException("Missing user")
     if new_user_individual.internal_id is None or new_user_individual.internal_id == "":
         raise PhenopolisException("Missing individual id")
+
+
+def _add_config_from_admin(db_session, new_users):
+    configs = db_session.query(UserConfig).filter(UserConfig.user_name.match("Admin")).all()
+    for u in new_users:
+        new_configs = []
+        for c in configs:
+            new_user_config = UserConfig(**c.as_dict())
+            new_user_config.user_name = u.user
+            new_configs.append(new_user_config)
+        db_session.add_all(new_configs)
